+++ src/ALCProxy.Communication/ServerDispatch.cs
@@ -3,34 +3,12 @@
 // See the LICENSE file in the project root for more information. 
 
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 using System.Runtime.Serialization;
 
 namespace ALCProxy.Communication
 {
-    public class ClientDispatch : ALCClient
-    {
-        public ClientDispatch(Type intType) : base(intType, "ServerDispatch`1") { }
-        protected override object SerializeParameter(object param, Type paramType)
-        {
-            MemoryStream stream = new MemoryStream();
-            //Serialize the Record object to a memory stream using DataContractSerializer.  
-            DataContractSerializer serializer = new DataContractSerializer(paramType);
-            serializer.WriteObject(stream, param);
-            return stream;
-
-        }
-        protected override object DeserializeReturnType(object s, Type returnType)
-        {
-            MemoryStream stream = (MemoryStream)s;
-            stream.Position = 0;
-            //Deserialize the Record object back into a new record object.  
-            DataContractSerializer newSerializer = new DataContractSerializer(returnType);
-            return newSerializer.ReadObject(stream);
-        }
-    }
     public class ServerDispatch<I> : ALCServer<I>
     {
         public ServerDispatch(Type instanceType, Type[] genericTypes, IList<object> constructorParams, IList<Type> constTypes) 
@@ -38,6 +16,10 @@ namespace ALCProxy.Communication
 
         protected override object DeserializeParameter(object serializedParam, Type t)
         {
+            if (!serializedParam.GetType().Equals(typeof(MemoryStream)))
+            {
+                throw new Exception("The Server passed the wrong type to the client when returning an object");
+            }
             MemoryStream s = (MemoryStream)serializedParam;
             s.Position = 0;
             //Deserialize the Record object back into a new record object.
